pyproject.toml
[tool.poetry]
name = "ceremo-services"
version = "0.1.0"
description = "Ceremo Services Flask Application"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
flask-sqlalchemy = "^3.0.0"
flask-migrate = "^4.0.0"
pyjwt = "^2.8.0"
bcrypt = "^4.1.0"
pydantic = {extras = ["email"], version = "^2.5.0"}
pymysql = "^1.1.0"
cryptography = "^41.0.0"
pytz = "^2024.1"
flask-cors = "^6.0.2"
tzdata = "^2025.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-flask = "^1.2.0"
pytest-mock = "^3.11.0"
pytest-cov = "^4.1.0"
responses = "^0.23.0"
black = "^23.12.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pre-commit = "^4.5.1"
types-pytz = "^2025.2.0.20251108"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests"
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
exclude = [
    "tests/",
    ".venv/",
]
